pyproject.toml
[project]
name = "conq_python"
version = "0.0.1"
description = "Python library and example scripts for Conq"
readme = "README.md" # Optional

requires-python = ">=3.8"

license = { file = "LICENSE.txt" }

keywords = ["robotics", "spot", "python"]
authors = [
    { name = "Peter Mitrano", email = "pmitrano@umich.edu" }
]
maintainers = [
    { name = "Peter Mitrano", email = "pmitrano@umich.edu" }
]

dependencies = [
    "Pillow",
    "bosdyn-choreography-client",
    "bosdyn-client",
    "bosdyn-mission",
    "matplotlib",
    "numpy",
    "opencv-python",
    "pip>=23.1.2",
    "scikit-learn",
    "scipy",
    "sklearn",
]

[project.optional-dependencies] # Optional
regrasping_demo = [
    "cvxpy",
    "cvxpylayers",
    "plotly",
    "pytorch_kinematics",
    "pytorch_volumetric",
    "requests",
    "rerun-sdk",
    "roboflow",
    "torch",
]

[tool.setuptools]

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"
